fix the amount for the admin

+++ database/migrations/2025_05_16_010857_create_admins_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('password');
             $table->string('email');
-            $table->integer('amount')->default(0);
+            $table->double('amount')->default(0);
             $table->timestamps();
         });
     }
